+++ ipwhois/tests/online/test_net.py
@@ -1,22 +1,12 @@
-import unittest
 import logging
+from ipwhois.tests import TestCommon
 from ipwhois import (Net, ASNLookupError, ASNRegistryError, BlacklistError,
-                     WhoisLookupError, HTTPLookupError, HostLookupError)
+                     WhoisLookupError, HTTPLookupError, HostLookupError,
+                     HTTPRateLimitError)
 
 LOG_FORMAT = ('[%(asctime)s] [%(levelname)s] [%(filename)s:%(lineno)s] '
               '[%(funcName)s()] %(message)s')
 logging.basicConfig(level=logging.DEBUG, format=LOG_FORMAT)
-
-
-class TestCommon(unittest.TestCase):
-
-    if not hasattr(unittest.TestCase, 'assertIsInstance'):
-        def assertIsInstance(self, obj, cls, msg=None):
-            if not isinstance(obj, cls):
-                self.fail(self._formatMessage(
-                    msg,
-                    '%s is not an instance of %r' % (repr(obj), cls)
-                ))
 
 
 class TestNet(TestCommon):
@@ -30,7 +20,7 @@ class TestNet(TestCommon):
         except AssertionError as e:
             raise e
         except Exception as e:
-            self.fail('Unexpected exception raised: %r' % e)
+            self.fail('Unexpected exception raised: {0}'.format(e))
 
     def test_get_asn_whois(self):
         result = Net('74.125.225.229')
@@ -41,7 +31,7 @@ class TestNet(TestCommon):
         except AssertionError as e:
             raise e
         except Exception as e:
-            self.fail('Unexpected exception raised: %r' % e)
+            self.fail('Unexpected exception raised: {0}'.format(e))
 
         result = Net('74.125.225.229')
         self.assertRaises(ASNLookupError, result.get_asn_whois, 3, 'a')
@@ -55,7 +45,7 @@ class TestNet(TestCommon):
         except AssertionError as e:
             raise e
         except Exception as e:
-            self.fail('Unexpected exception raised: %r' % e)
+            self.fail('Unexpected exception raised: {0}'.format(e))
 
         self.assertRaises(WhoisLookupError, result.get_whois, **dict(
             retry_count=0, server='arin.net'))
@@ -78,14 +68,35 @@ class TestNet(TestCommon):
         except AssertionError as e:
             raise e
         except Exception as e:
-            self.fail('Unexpected exception raised: %r' % e)
+            self.fail('Unexpected exception raised: {0}'.format(e))
 
         self.assertRaises(HTTPLookupError, result.get_http_json, **dict(
             url='http://255.255.255.255', retry_count=0))
 
         result = Net('74.125.225.229', 0)
+        url = RIR_RDAP['arin']['ip_url'].format('74.125.225.229')
         self.assertRaises(HTTPLookupError, result.get_http_json, **dict(
-            retry_count=1))
+            url=url, retry_count=0))
+
+        # Uncommenting below will result in a flood of up to 20 requests
+        # to test rate limiting.
+        '''
+        url = RIR_RDAP['lacnic']['ip_url'].format('200.57.141.161')
+        result = Net('200.57.141.161')
+        count = 20
+        while count > 0:
+            count -= 1
+            try:
+                self.assertRaises(HTTPRateLimitError, result.get_http_json,
+                                  **dict(url=url, retry_count=0))
+                break
+
+            except AssertionError as e:
+                if count == 0:
+                    raise e
+                else:
+                    pass
+        '''
 
     def test_get_host(self):
         ips = [
@@ -102,7 +113,7 @@ class TestNet(TestCommon):
             except AssertionError as e:
                 raise e
             except Exception as e:
-                self.fail('Unexpected exception raised: %r' % e)
+                self.fail('Unexpected exception raised: {0}'.format(e))
 
         result = Net('74.125.225.229', 0)
         self.assertRaises(HostLookupError, result.get_host, **dict(
@@ -117,7 +128,12 @@ class TestNet(TestCommon):
         except AssertionError as e:
             raise e
         except Exception as e:
-            self.fail('Unexpected exception raised: %r' % e)
+            self.fail('Unexpected exception raised: {0}'.format(e))
 
         result = Net('74.125.225.229', timeout=0, allow_permutations=False)
         self.assertRaises(ASNRegistryError, result.lookup_asn)
+
+        result = Net('74.125.225.229', timeout=0, allow_permutations=True)
+        self.assertRaises(HTTPLookupError, result.lookup_asn, **dict(
+            asn_alts=['http']
+        ))